+++ src/server/services/utils/utils.ts
@@ -300,4 +300,27 @@ export class Utils {
     }
     return hex;
   }
+
+  public static hexToRgb(color: string): { r: number; g: number; b: number } | null {
+    let hex: string | null = Utils.formatHex(color);
+    if (hex === null) {
+      return null;
+    }
+    const shorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])$/i;
+    hex = hex.replace(shorthandRegex, (_m, r, g, b) => {
+      return r + r + g + g + b + b;
+    });
+
+    const result: RegExpExecArray | null = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+    if (!result || result.length !== 4) {
+      return null;
+    }
+    const r: number = parseInt(result[1], 16);
+    const g: number = parseInt(result[2], 16);
+    const b: number = parseInt(result[3], 16);
+    if (isNaN(r) || isNaN(g) || isNaN(b)) {
+      return null;
+    }
+    return { r, g, b };
+  }
 }