+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Exchange Govee-Lan-Controller
+
 ## 2.22.9 (2024-01-05)
 
 * (theimo1221) Extend Logging for all API calls